requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
openai==1.12.0
twilio==8.12.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
airtable-python-wrapper==0.15.3
pytest==8.0.0
httpx==0.26.0
python-multipart==0.0.9
alembic==1.13.1
redis==5.0.1
tenacity==8.2.3
streamlit==1.31.1
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client 